Heading

I started working with Rust in 2020 at RustMinded, where I built a strong foundation in the language and the open-source ecosystem. Since then, I’ve contributed to WebAssembly and Rust tools, worked on rewriting a data query engine to improve performance, or helped on the development of a cybersecurity platform for analyzing vulnerabilities on multiple systems. Along the way, I’ve gained practical experience with JavaScript and TypeScript, working in team environments to deliver effective solutions. I’m also an active contributor to open-source projects and rely on Linux to streamline my development workflow.

You can find me on GitHub and Bluesky. You can also send me an email.

Subtitle

Element

Freelance mission as a Rust Developer at Ystorian working on a cybersecurity platform that gathers information from the OS, BIOS, kernel or network on different systems in parallel to assess security vulnerabilities and compliance level.

Element

Freelance mission as a Rust Developer at Cumul.io rewriting in Rust a query engine for data analytics using faster and more efficient queries and improving the relation of multiple microservices in an AWS cloud environment.

Element

First job with hands on the keyboard at RustMinded. Discovering the OSS world when learning how to become an Rustacean. Now working as Rust Developer Consultant.

Element

Multiple jobs with hands on the wheel (emergency response driver, airport shuttle, taxi,…).

Subtitle

Element

A CLI tool that allows you to create a temporary new Rust project using cargo with already installed dependencies.

Element

A library that provides an API similar to std::process to create and handle processes on Windows using the Win32 API.

Element

A Yew and Yewprint playground using wasm-run.

Subtitle

Element

A developer tool to help you get up and running quickly with a new Rust project by leveraging a pre-existing git repository as a template.

Element

Date and Time for Rust.

Element

A GPT manager that allows you to copy partitions from one disk to another.

Element

Immutable types and ImplicitClone trait similar to Copy.

Element

French translation of the book “The Rust Programming Language”.

Element

A library to get system information such as processes, Cpus, disks, components and networks.

Element

Third-I, the only camera that replicates your hearing and your point of view.

Element

TopoJSON bindings and utilities for Rust.

Element

Customizable commands based on xtask for your WASM project.

Element

A customizable helper to watch for changes in your xtask projects.

Element

A port of blueprintjs to Yew.

Subtitle

Element

Release, update and maintenance of an OSS project.

Element

Project documentation using MarkDown, Mermaid, rustdoc,…

Element

Project organisation using GitHub Project and GitHub settings.

Element

Continuous Integration via GitHub Actions.

Element

Web application deployment using Netlify.

Subtitle

Element

Yew, Rocket, Axum, Actix, Tokio, sqlx.

Element

Nixos, Arch Linux, Pop!_OS, macOS, Windows, Debian, Tails.

Element

Rust, Bash, HTML, CSS, JavaScript.

Element

French (5. Native), English (3. Minimum Professional Proficiency), Dutch (1. Basic).

Element

Git, WebAssembly, CLI, Terminal UI, Platform agnosticism, System, Web, Backend, Containerization, Developer Tooling.

Subtitle

Element

Element

Element

Element

Element

Element

Element