Crate yohan_boogaert

A former emergency response driver who always been interested in programming and technology. I’ve been running Linux for 6 years, I tried many distributions (Ubuntu, Tails, Manjaro,…). I use Arch Linux as my daily driver because Arch gives me the opportunity to really discover what it’s going under the hood and nearly build my desktop from scratch.

I heard about Rust in 2020 from a friend who shared her knowledge with me about programming and helped me with Linux. Recently that friend (who is my mentor now) gave me the opportunity to realize my dreams and become a developer! I now learn how to write Rust code and how to work like a developer at RustMinded.

You can find me on Github and Twitter. You can also send me a mail.

Modules

emergency_medical_technician
Volunteer at Centre de Secours Monceau - 2020

professional_driver
Ambulance, Shuttle, Taxi,… - 2014 to 2019

software_developer
Junior at RustMinded - Current

Structs

CargoTemp
Project - A CLI tool that allows you to create a temporary new Rust project using cargo with already installed dependencies. site

IntroToRust
Project - My vision of the Rust Programming Language as an introduction to Rust. site

YewprintPlayground
Project - A Yew and Yewprint playground using wasmbl. site

Enums

CargoBreaking
Contributions - A CLI tool that compares a crate’s public API between two different branches, shows what changed, and suggests the next version according to semver. site

RustBookFr
Contributions - French translation of the book “The Rust Programming Language”. site

ThirdIFrontend
Contributions - “Third-I”, the only camera that replicates your hearing and your point ofview. site

Wasmbl
Contributions - Bundles and tooling for Rust WASM frontend application. site

Yewprint
Contributions - Port of blueprintjs.com to Yew. site

Constants

FRAMEWORKS
Yew.

OS
Arch Linux, Pop!_OS.

PROGRAMMING_LANGUAGES
Rust, Bash.

SPOKEN_LANGUAGES
French (5. Native), English (3. Minimum Professional Proficiency).

TECHNOLOGIES
Git, WASM, CLI.

Traits

Autodidact

Enthusiast

Flexible

Passionnate

TeamPlayer

WantToBuild

WantToLearn